+++ Adj/Algebra/Unit.hs
@@ -1,0 +1,3 @@
+module Adj.Algebra.Unit where
+
+type family Unit (p :: * -> * -> *) = r | r -> p